Make it easier to understand that 'properties' is similar to WHERE in SQL.
I think this is useful to note to implementers.  This also exposes a need to possibly add a new parameter for 'properties_strict' (like type_strict) where the value might be "must" or "should" and would offer a slightly different score?

I'm not sure if I wrote this up in entirely the best way, but I wanted to make it easier to understand since many folks know how database querying works.  Just the original word "refine" I think is not enough here.  Also, we should note places in the specs where clients can specify "conditions" for filtering and how scoring might be controlled or affected further, but unsure how best to present that.